Import models to own format and loading from the resources panel
Now you can import models from outside of the engine and by dragging the file to the engine, it is converted into an own format which is ready to use in the engine.

By now, only models with one mesh are supported.

Also some changes on the resources panel decoration.